Patched s3Client.destroy() in put-object.js

+++ nodes/put-object/put-object.js
@@ -298,7 +298,7 @@ function inputHandler(n, RED) {
       msg.key = n.key;
       send(msg);
     } finally {
-      s3Client.destroy();
+      if(s3Client) s3Client.destroy();
       /* Dereference vars */
       s3Client = null;
       /*********************/
